Crate waf_cargo_build

Waf Cargo binding library

This library helps binding the Waf build system and Cargo build system.

The Waf build system generates a .waf-cargo-build/waf_build_env.json to pass the environment or the package using this library. This library is used in build scripts.

The main entry point is the structure WafBuild.

Macros

include_bindings

Macro to include bindings.

Structs

WafBuild

Attribute Macros

bindings_mod

Macro to simplify the bindings module declaration.